fix: revise naming
This changes the naming and semantics of "evaluation results" to
"evaluation state".

We want to maintain a distinction between the two, where "evaluation
results" specifically refers to the boolean or string `result` resulting
from an evaluation based on the "result" of a top-level rule.

Whereas "evaluation state" is meant to refer to the current boolean
or null evaluation status of the overall ruleset and individual rules.

A null evaluation status indicates the ruleset (and/or rule) has yet to
be evaluated.